+++ deployment/ccip/changeset/solana_v0_1_1/testutils_test.go
@@ -1,5 +1,3 @@
-//go:build integration && solana
-
 package solana_test
 
 import (
