+++ deployment/ccip/shared/stateview/view_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && db
+
 package stateview_test
 
 import (
