+++ deployment/internal/jdtestutils/client_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && db
+
 package jdtestutils
 
 import (
