+++ deployment/utils/nodetestutils/node_integ_test.go
@@ -1,3 +1,5 @@
+//go:build integration && db
+
 package nodetestutils
 
 import (
